+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this library are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fluentassert/verify/compare/v1.0.0...HEAD)
+## [Unreleased](https://github.com/fluentassert/verify/compare/v1.1.0...HEAD)
+
+## [1.1.0](https://github.com/fluentassert/verify/releases/tag/v1.1.0) - 2024-02-06
+
+This release adds length assertions.
+
+### Added
+
+- Add `Len` assertion for `string`, `[]T`, `map[K]V` types.
 
 ## [1.0.0](https://github.com/fluentassert/verify/releases/tag/v1.0.0) - 2023-04-05